requirements.txt
streamlit
google-cloud-storage
google-cloud-aiplatform